remove VNC_PW from environment to protect privacy;

+++ src/common/scripts/vnc_startup.sh
@@ -71,7 +71,7 @@ echo -e "\n------------------ change VNC password  ------------------"
 mkdir -p "$HOME/.vnc"
 PASSWD_PATH="$HOME/.vnc/passwd"
 
-if [[ -f $PASSWD_PATH ]]; then
+if [[ -f $PASSWD_PATH && -n ${VNC_PW} ]]; then
     echo -e "\n---------  purging existing VNC password settings  ---------"
     rm -f $PASSWD_PATH
 fi
@@ -81,7 +81,13 @@ if [[ $VNC_VIEW_ONLY == "true" ]]; then
     #create random pw to prevent access
     echo $(head /dev/urandom | tr -dc A-Za-z0-9 | head -c 20) | vncpasswd -f > $PASSWD_PATH
 fi
-echo "$VNC_PW" | vncpasswd -f >> $PASSWD_PATH
+if [ ! -f $PASSWD_PATH ]; then
+	if [ -z ${VNC_PW} ]; then
+		echo "error: no '$PASSWD_PATH' exist or 'VNC_PW' provided"
+		exit 1
+	fi
+	echo "$VNC_PW" | vncpasswd -f >> $PASSWD_PATH
+fi
 chmod 600 $PASSWD_PATH
